rename sblogPost list to index.js
